requirements.txt
Django
Pillow
google-generativeai
python-decouple
